requirements.txt
# Content Creation Multi-Agent System Dependencies
# AAIDC Module 2 Project - Local Ollama Version

# Core LangChain and LangGraph dependencies (compatible versions)
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langgraph>=0.0.55
langchain-core>=0.2.20

# Ollama integration
ollama>=0.2.1

# Web search and scraping tools
duckduckgo-search==5.3.0
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7

# Natural Language Processing
nltk==3.8.1
textstat==0.7.3
spacy==3.7.4

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3

# Async support
aiohttp==3.9.5
asyncio-throttle==1.0.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.7.1

# Logging and monitoring
structlog==23.1.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.7.0
flake8==6.0.0

# Optional: For advanced features
# google-search-results==2.4.2  # For Google Search API
# grammarly-api==0.1.0  # For grammar checking (if available)
# readability==0.3.1  # Alternative readability calculator